moving to Point3<float> mathematics, lot of testing needed

+++ Sources/Common/Game/Position.hpp
@@ -7,7 +7,7 @@ namespace Common
 namespace Game
 {
 
-typedef Common::Math::Point3<int> Position;
+typedef Common::Math::Point3<float> Position;
 
 }
 }
